Revert "Typescript backend"
