+++ .github/workflows/deploy-new.yml
@@ -89,7 +89,6 @@ jobs:
         run: |
           pnpm install --no-frozen-lockfile
           pnpm run build
-          cd ./server/ && pnpm run build
         env:
           CI: false
 
@@ -144,7 +143,7 @@ jobs:
 
             echo -e "SECRET=${{ secrets.SECRET }}\nDB_USER=${{ secrets.DB_USER }}\nDB_PASS=${{ secrets.DB_PASS }}\nDB_AUTH=${{ secrets.DB_AUTH }}\nDB_NAME=${{ secrets.DB_NAME }}\nPORT=${{ secrets.PORT }}\nGOOGLE_CLIENT_ID=${{ secrets.GOOGLE_CLIENT_ID }}\nGOOGLE_CLIENT_SECRET=${{ secrets.GOOGLE_CLIENT_SECRET }}" > .env
 
-            pnpm install --frozen-lockfile
+            pnpm install --frozen-lockfile --ignore-scripts
 
             if [ "${NEW_CLONE}" = true ]; then
               pm2 start ecosystem.config.cjs
