+++ server/ecosystem.config.cjs
@@ -2,8 +2,8 @@ module.exports = {
 	apps: [
 		{
 			name: "rebound-express",
-			script: "./build/app.js",
-			watch: ["build", ".env", "package.json", "pnpm-lock.json"],
+			script: "./src/app.js",
+			watch: ["src", ".env", "package.json", "package-lock.json"],
 			watch_delay: 250,
 			ignore_watch: ["dev", "node_modules", "uploads"],
 			exp_backoff_restart_delay: 100,
