+++ server/jsconfig.json
@@ -1,0 +1,8 @@
+{
+	"compilerOptions": {
+		"baseUrl": "./src",
+		"target": "es6",
+		"module": "commonjs"
+	},
+	"exclude": ["node_modules"]
+}
