+++ server/src/app.js
@@ -16,22 +16,23 @@ import socketBackend from "./socketio/index.js";
 configDotenv();
 
 class ServerBackend {
-	public app: express.Express;
-	private server: http.Server;
-
 	constructor() {
 		this.app = express();
-		new customPassport().setupPassport();
+		customPassport.setupPassport();
 		this._initMiddleware();
 		this._initRoutes();
 		this.server = http.createServer(this.app);
 	}
 
 	_initMiddleware() {
-		this.app.set("trust proxy", 1); // trust nginx proxy
-		this.app.use(cors({ optionsSuccessStatus: 200 }));
+		// Trust the reverse proxy (e.g. Nginx) when determining
+		// protocol and other forwarding headers
+		this.app.set("trust proxy", 1);
 
-		// GLOBAL RATE LIMITER
+		// CORS
+		this.app.use(cors({ optionsSuccessStatus: 200 }));
+		// ─── GLOBAL RATE LIMITER ───────────────────────────────────────────────────
+		// limit each IP to 150 requests per 5 minutes
 		const globalLimiter = rateLimit({
 			windowMs: 5 * 60 * 1000, // 5 minutes
 			max: 150,
