--- server/tsconfig.json
@@ -1,17 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es2020",
-		"module": "es2020",
-		"moduleResolution": "node",
-		"outDir": "build",
-		"rootDir": "src",
-		"strict": true,
-		"noImplicitAny": false,
-		"allowJs": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"skipLibCheck": true
-	},
-	"include": ["src/**/*"],
-	"exclude": ["node_modules"]
-}
